firmware_TouchpadMTB: add a scrolled window for result frame
Some test result texts may be too long or too wide such that they
are cut off at the end of the window. In this patch, we add a
scrolled window to the result frame so that the user can use the
arrow keys to scroll the test result texts around.

BUG=chromium:216688
TEST=perform the firmware tests on a chromebook:
$ cd /usr/local/autotest/tests/firmware_TouchpadMTB
$ DISPLAY=:0 OPTIONS=-s python main.py
Observe that when the test results are too long, the corresponding
scroll bars show up. You could press UP, DOWN, LEFT, or RIGHT arrow
keys to scroll the result texts.

Change-Id: I175dcb3c4952055a534a1c5b01811b23b0b649bf
Reviewed-on: https://gerrit.chromium.org/gerrit/47848
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>