[Autotest] Pylint checks all files if the only change is to remove a file.
As an artifact of the old run_pylint we maintained the option of checking all
files in a directory if the list of presubmit files is empty. This allows one
to check all of autotest by invoking ./utils/run_pylint.py. We should instead
only do this if the PRESUBMIT_FILES environment variable is not set. If the
list is empty we could still be removing files, and we don't need to check them.

TEST=Ran repo upload . with the offending cl and made sure pylint didn't
     complain.
BUG=None

Change-Id: Ie38b7a2a28b89677a56b0f36f06981aaedf7be3d
Reviewed-on: https://gerrit.chromium.org/gerrit/47912
Reviewed-by: Tan Gao <tgao@chromium.org>
Commit-Queue: Prashanth Balasubramanian <beeps@chromium.org>
Tested-by: Prashanth Balasubramanian <beeps@chromium.org>