Fix bugs in nmo testing.
After adding logs I found two issues that are related to my changes
to modifying the dev_server code. This fixes the url returned
in dev_server.py for delta payloads to be correct.

I've also added additional logging to make this easier to find in the
future.

FYI: The reason why this was only sometimes failing was that the delta
payload url being returned ended up being the full payload url. So if
the image was staged on the same devserver as was chosen for the
devserver delta payload, it would pass, otherwise it would fail.

BUG=chromium:226159
TEST=Verified on a local dut through the test.

Change-Id: Ib3f00dc43f806f05d500062972b506954b8b7002
Reviewed-on: https://gerrit.chromium.org/gerrit/47914
Reviewed-by: Chris Sosa <sosa@chromium.org>
Tested-by: Chris Sosa <sosa@chromium.org>