Change NAME attribute in dynamically generated control files to include suffix.
Dynamic suite uses the name of the test in the control file to find test logs.
Currently when we run the au suite, run_suite cannot find the right test log
because it returns just the first matching test. See:
crbug.com/226147. This fixes that by re-encoding the NAME when we generate the
file.

BUG=chromium:226147
TEST=Ran it locally and verified names were correct.

Change-Id: Icd2a74bd134dbabb371e4a6b600d699d032f4f80
Reviewed-on: https://gerrit.chromium.org/gerrit/47934
Tested-by: Chris Sosa <sosa@chromium.org>
Reviewed-by: Gilad Arnold <garnold@chromium.org>
Commit-Queue: Chris Sosa <sosa@chromium.org>