Add Z-calibration to the Touch FW test
This patch adds in a chunk of code to run the z-calibration script
automatically whenever you are using the robot. So the first thing that
happens now if you are running the robot part of the test is that it
attempts to determine the Z dimension of the touchpad and uses that
instead of the passed-in values

BUG=chromium:229600
TEST=Manually tested on a Lumpy.  It sucessfully calibrated and ran the
tests with that value.

Change-Id: I71f38e94b308a41339797591dd1640bc4dc043fe
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/48169