Make suspend tests wait for syslog flush before parsing it
The current suspend test code can raise with rsyslogd writing back its
log files to the disk. There are a few retries involved that usually
catch the worst of it, but it might still be possible to miss a warning.
This patch solves that issue by synchronously waiting until the reliable
final message from powerd_suspend appears in the log. This also makes
the sanity check whether the system actually tried to suspend a little
more elegant.

Additionally, this solves a minor issue where a log line could be
chopped in half between reads. It also upgrades all synchronous waiting
behavior to use an initially very short, exponentially growing sleep
time, in order to make the common case of short delays as fast and
responsive as possible.

BUG=chromium:231102
TEST=power_Resume, power_UiResume, power_SuspendStress

Change-Id: I154e87cadc494d8f67aff329f8c472c74d114d66
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/48259
Reviewed-by: Mark Koudritsky <kamrik@chromium.org>