Update the stateful partition during an update test.
For test images, we may need to update the stateful partition during
a test as Y may not be compatible with X. For example, we just got
burned with the Python2.6->Python2.7 upgrade where the new rootfs
symlinked /usr/bin/python to /usr/local/bin/python2.7 which doesn't
exist unless you update the stateful partition as well.

As part of this, I have exposed the stateful update from autoupdater
to make it easier to use. I have also added unittests for this method.

BUG=chromium:232085
TEST=pylint + manual

Change-Id: I6088cf671d090fca7aabff25335f4c0b8e3947c1
Reviewed-on: https://gerrit.chromium.org/gerrit/48331
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>
Tested-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Chris Sosa <sosa@chromium.org>