firmware_TouchpadMTB: fix a bug about logs for multiple iterations
When we use the command line options such as "-i 3" to record 3
iterations continuously, the test result log should keep all the
scores of the 3 iterations. The firmware_summary module should be
able to parse the score list in the log correctly.

BUG=chromium:234068
TEST=Perform the following steps on a chromebook machine:
$ cd /usr/local/autotest/tests/firmware_TouchpadMTB
$ python tests/firmware_summary_unittest.py
The unit tests should all pass.

Change-Id: I8332ef1881de801cd6ebd7efbf344cb0fc25d545
Reviewed-on: https://gerrit.chromium.org/gerrit/48780
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>