Autotest job tags in server-side suites are passed to underlying client jobs.
This change reverts one part of:

https://gerrit.chromium.org/gerrit/#/c/41318/4

Previously, when a server-side suite invoked a client-side test with
a tag, that tag would be passed to the client job and appended to the
name of the client-side test.  For example, control.power_daily
invokes power_LoadTest multiple times with different tags.  This
would lead to tests with different names, such as "power_LoadTest.WIFI_1hr"
and "power_LoadTest.WIRED_1hr".

After the above CL was submitted, the tags were instead passed to the
server job, not to the client jobs.  This resulted in the tags _not_
being appended to the client-side test names, meaning that we could
no longer use the test names (with tags) to distinguish test results
from different invocations of the same client-side test.

In the current change, that tags are instead passed to the underlying
client jobs as before, so the tag can be appended to the names of the
invoked client-side tests.

BUG=chromium:227172
TEST=Modified control.power_daily to invoke the "sleeptest"
client-side test with different tags, and verified that the resulting
test names _do_ include the tag with this change, but don't include
the tags without this change.

Change-Id: I44581ea84e990995aa6323e9cb517419b5e3c70e
Reviewed-on: https://gerrit.chromium.org/gerrit/48563
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Queue: Dennis Jeffrey <dennisjeffrey@chromium.org>
Tested-by: Dennis Jeffrey <dennisjeffrey@chromium.org>