[autotest] Allow passing a connection arg to stats utility constructors
Currently get_client of the stats utility classes doesn't work.
For example, site_utils.graphite.stats.Timer.get_client raises
an error. This is because get_client in its base class tries to construct
an instance of the current class with args |name| and |connection|.
However, the |__init__| method of site_utils.graphite.stats.Timer
does not accept arg |connect|. This CL fixes the problem
by adding a connection arg to constructors of stats utility classes.
Note that if a |connection| is passed, it will
set bare to True thus will not alter the name for you.

BUG=None
TEST=run get_client and it works.

Change-Id: Id5d67a176eef46fa7e42bd8dd4ff981f324b363a
Reviewed-on: https://gerrit.chromium.org/gerrit/48583
Reviewed-by: Alex Miller <milleral@chromium.org>
Tested-by: Fang Deng <fdeng@chromium.org>
Commit-Queue: Fang Deng <fdeng@chromium.org>