Fix power_SuspendStress.idle timing problems
This patch fixes several timing issues in the idle suspend stress test
that have always been there but have been exacerbated with CL:48259. It
increases the synchronization delay that power_SuspendStress waits on
dummy_IdleSuspend even more, because PyAuto just takes that damn long to
do a simple login. It reduces the maximum time to wait for the
hwclock-on-resume to appear to ~300ms, as that is intended to fail in
the Lumpy/Stumpy RTC bug case and our new design already waits on
powerd_suspend messages before that anyway (so there should really be no
cases in practice where we need to wait on that file at all). Finally,
it increases the test's minimum suspend duration a bit just to add a
safety buffer and eliminates the breathing_time to reduce the wait
after resume until the RTC is scheduled again (we don't need it here
anyway, since the system always stays active for 8s while waiting for
the idle timeout which should be plenty of time for network activity).

BUG=chromium:220217,chromium:234224
TEST=None

Change-Id: I41238fa4c32bb80dba80b821f94e9fc58f4bf534
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/48811
Reviewed-by: Alex Miller <milleral@chromium.org>