[autotest] Pylint shouldn't fail with empty file paths.
In order to control exceptions, messages and pylint exiting, we
invoke the Run method on pylint directly instead of shelling out
and calling the module itself. This Run method takes the command
line arguments and a list of files to parse. Many of these
command line arguments trigger the execution of a callback
function that will simply exit when the list of files to check
is empty. Normal invocations of pylint are not expected to reach
this stage, but we do because of the direct invocation mentioned.

TEST=Ran repo upload . with the offending commit which contains a
     list of non-py files and made sure we don't exit.
BUG=None

Change-Id: Id2ea1ac8e025e235c35d1ccf08859ead1d6d7cfb
Reviewed-on: https://gerrit.chromium.org/gerrit/48810
Reviewed-by: Alex Miller <milleral@chromium.org>
Commit-Queue: Prashanth Balasubramanian <beeps@chromium.org>
Tested-by: Prashanth Balasubramanian <beeps@chromium.org>