[autotest] Reland Aborting a job aborts all child jobs.
Django has been upgraded to 1.5, so this is now safe to land.

When a HQE is aborted, we find all HQEs associated with a parent_id of
job that is being aborted, and abort those also.

This was implemented as "requesting a job to abort also requests all
child jobs to abort".  All calls to abort a host queue entry end up in
the afe model's abort().  If one directly modifies the database, one can
still abort a singular job without also aborting its children, in case
anyone ever has a need to do so.

BUG=chromium:226826
DEPLOY=apache, scheduler
TEST=Kicked off two suites.  Aborted first one immediately, and the
other suite was unaffected.  Aborted the second suite after the first
job finished, and the completed job was unaffected.  All other tests
(and suite job) were aborted.  Confirmed in database that the host queue
entries rows were set correctly, and the appropriate rows in the
|afe_aborted_host_queue_entries| were still being inserted.
Modified a suite after it was made so that it was its own parent.
Confirmed that aborting the suite ran correctly.

Change-Id: I48303dfe5264d24183c66fa7fd2158a6b838cd51
Reviewed-on: https://gerrit.chromium.org/gerrit/48864
Tested-by: Alex Miller <milleral@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Alex Miller <milleral@chromium.org>