[autotest] Devserver resolution shouldn't fail unless all devservers are down.
In trying to resolve a build to a devserver instance we go through the list of
devservers looking for one to use. This leads to a devserer_up call which
checks to see if it can access the index of the devserver, and waits for about
six seconds for a reply. If this call throws a TimeoutException we fail, we
should instead catch it and move on to the next devserver since if it's a real
devserver problem Nagios will send out mail.

TEST=Called the method in both the pass and fail case and checked behavior.
BUG=chromium:235164

Change-Id: Id9c52f3fdf47d464b6f9e77355e5049b1dc46d44
Reviewed-on: https://gerrit.chromium.org/gerrit/49088
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Prashanth Balasubramanian <beeps@chromium.org>
Tested-by: Prashanth Balasubramanian <beeps@chromium.org>