Web_driver_helper and Western_ditial configurator
In web_driver_core_helper we will now raise an
exception if we don't find any of the xpath's.
In Western Digital, we now wait for the page to
reload so that the element is in the cache when
we look for it and we can toggle visibility.

TEST=None
BUG=chromium:235139, chromium:223755, chromium:232212

Change-Id: I685810f50bfc7b85d0f0f9969102c9f4381c79b1
Reviewed-on: https://gerrit.chromium.org/gerrit/49137
Commit-Queue: Deepak Gopal <deepakg@chromium.org>
Reviewed-by: Deepak Gopal <deepakg@chromium.org>
Tested-by: Deepak Gopal <deepakg@chromium.org>