Check if is_visibility_supported before setting
To reduce assumptions and make test robust, if a configurator
does not support setting visibility and set_visibility is called
an exception will be thrown.

If set_visibility is not supported that means it is always broadcasting
the SSID.  For most tests this is a no-op.  However when we write tests
for SSIDs we will want to only use configurators that support it.

TEST=Manual, ran it against the Buffalo wzr_f1800h
BUG=chromium:235265

Change-Id: I5d0f6007725870c1e62d92b0e24a59839500f453
Reviewed-on: https://gerrit.chromium.org/gerrit/49129
Tested-by: Kris Rambish <krisr@chromium.org>
Reviewed-by: Tan Gao <tgao@chromium.org>