autotest: Extract WiFiContext server/router address lookup into properties
This allows subclasses of WiFiTestContextManager to change how these
lookups are performed.  For instance, RvR tests can only be run in
cells equipped with an attenuator, and we have few enough of such cells
that we simply hardcode the addresses.

BUG=None
TEST=Ran:
./run_remote_tests.sh --remote=chromeos1-shelf1-host3 check11b
and noted that host lookup works as expected for lab DUTS.

Change-Id: Icdf4d293866bf119921608bd8fb8ef8195a1f9fe
Reviewed-on: https://gerrit.chromium.org/gerrit/49342
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>