[Autotest] Remove empty space and tab in image url and version
scheduler may pass in an image option with \t at the end in autoserv command
line. The causes autoupdater failed to match the version from image or
update_version with lsb-release version string. The fix is to filter out any
tab and empty spaces in the beginning and end of image option for autoserv, as
well as in method autoupdater.url_to_version.

BUG=chromium:236382
TEST=local manual test
DEPLOY=none

Change-Id: Ideae45cd46c9cc34522fdf21aabaf067be9e6504
Reviewed-on: https://gerrit.chromium.org/gerrit/49495
Tested-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Queue: Dan Shi <dshi@chromium.org>