autotest: Extract more generic xmlrpc functionality from WiFi code
Both the dbus_safe decorator and the trivial ready() method are
generic and useful for writing your own xmlrpc delegate.  Put them next
to the XmlRpcServer where they'll be see and potentially used.

BUG=None
TEST=Ran:
./run_remote_tests.sh --remote=chromeos1-shelf1-host3 check11b
which constructs an XMLRPC server on the DUT and calls into it.

Change-Id: Ib43513fd5f0525475f414d8cc60e120476bb579e
Reviewed-on: https://gerrit.chromium.org/gerrit/49500
Commit-Queue: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Tested-by: Christopher Wiley <wiley@chromium.org>