Avoid spamming a hotplug event every time we create a Servo object.
When we create a Servo object, we set the USB mux switch to point to
the servo host.  The switching code didn't take account of the
current physical setting of the mux, with the result that we created
a hotplug event even when the mux setting didn't change.

An ongoing (and unexplained) failure on beaglebone is triggered by
hotplug events.  To reduce the frequency of these failures (and in
general, just to be nice), we change the logic to test the hardware
setting, and only change the mux setting when it's actually needed.

BUG=chromium:225932
TEST=platform_CloseOpenLid with the mux set both ways

Change-Id: Id3c6a05c5ae9905f12dfaa0d808b121401afe03a
Reviewed-on: https://gerrit.chromium.org/gerrit/49513
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>