[autotest] Index the query for special tasks for the scheduler.
This query takes ~5 seconds to run, scans ~100,000 rows, and is
unindexed.  This alone is about 45% of the scheduler tick.

Finding what indexes to create was done by taking the django query that's built
up, pasting it into a python interpreter, and calling
`queued_tasks.query.get_compiler('default').as_sql()` on it to get the SQL
Django generates.  One can then run `EXPLAIN SELECT ... \G` to get MySQL to
show how many rows it looks at for each part of the query, and if it's using an
index to do so or not.  In this case, we were scanning huge numbers of rows in
both `afe_special_tasks` and `afe_host_queue_entries`.

One can then create indexes for the columns that are being looked in the query.
In our case, trying to find special tasks that haven't been run yet required a
table scan to look at the `is_active` and `is_complete` columns, so those get
an index.  The query that looks at `afe_host_queue_entries` looks at `host_id`
and `active` at the same time, so we need to index both of those together too.

On a local prod-like setup, these indexes took the query from 30s down to 0.03s.

BUG=chromium:236636
DEPLOY=migrate
TEST=Ran migration, ran queries.

Change-Id: I86fb21716f7aa2cb7e04a4f461581a64be2e9652
Reviewed-on: https://gerrit.chromium.org/gerrit/49711
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Queue: Alex Miller <milleral@chromium.org>
Reviewed-by: Alex Miller <milleral@chromium.org>
Tested-by: Alex Miller <milleral@chromium.org>