Stage test images for servo automatically.
Instead of using a hard-coded URL that must be manually updated,
stage test images on the devserver on demand for servo repair.  The
version to stage is selected by the "stable_cros_version" global
config setting; this is the same setting used when repairing via
machine_install().

BUG=chromium:215430
TEST=repair in a local autotest instance, plus platform_InstallTestImage

Change-Id: Id7cefa6fedf3ad0033f693b3e4e0ce231eda9d63
Reviewed-on: https://gerrit.chromium.org/gerrit/49778
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>