Autotest: run run_cros_tests in the lab.
We now have a subsection of chrome-os specific telemetry tests. These
tests are kicked off by a different script. Thus the autotest
telemetry runner code needs to be aware of its existence and allow
access to it.

Also added an example test so that people know how to add tests in
the future.

BUG=chromium:234839
TEST=Ran the new test.

Change-Id: I7bf81f1c9eb03aa5af45e85bfbe3252bad5d3aff
Reviewed-on: https://gerrit.chromium.org/gerrit/49908
Commit-Queue: Simran Basi <sbasi@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>