Perf dashboard: do not filter out perf values that are 0.
A bug exists in the script that extracts perf data from the autotest
results database to display on the perf dashboard.  The effect of the
bug is that all perf keyvals that have a value of 0 are omitted from
being displayed on the perf dashboard.  This is now fixed.

BUG=chromium:237171
TEST=Verified that perf keyvals previously missing from
platform_GesturesRegressionTest because they had values of 0,
are now properly displayed on a local instance of the perf
dashboard.

Change-Id: Icd6aebacdeaa84b303959a9d94c68f2671349388
Reviewed-on: https://gerrit.chromium.org/gerrit/49957
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Queue: Dennis Jeffrey <dennisjeffrey@chromium.org>
Tested-by: Dennis Jeffrey <dennisjeffrey@chromium.org>