Omit logging the hwid after a test run if it's not available.
After an autotest run, we invoke "crossystem hwid" in order to log
the hwid of the device on which the test ran, as a test keyval.
This value may not always be available (e.g., when running on a VM).

As a result, if "crossystem hwid" fails, we use the output of
a second command -- "crossystem mainfw_type" -- to determine whether
or not we expect "crossystem hwid" to be available.  If the output
of that second command is "nonchrome", then we expect the hwid to
not be available, and we proceed.  Otherwise, the hwid is unexpectedly
missing, and we raise an exception.

BUG=chromium:223728
TEST=Verified on a local test run that if the "crossystem hwid" command
fails, we now either proceed or else raise an exception, depending on
the output of "crossystem mainfw_type".

Change-Id: I52b8e52efe645b20ef558ba0c97326d710c89db5
Reviewed-on: https://gerrit.chromium.org/gerrit/50067
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Queue: Dennis Jeffrey <dennisjeffrey@chromium.org>
Tested-by: Dennis Jeffrey <dennisjeffrey@chromium.org>