autotest: Convert DTIM period test to new WiFi framework
This is another straightforward migration of a test to the new WiFi
framework.  As part of this change, we migrate powersave controls to
WiFiClient, and add DTIM period support to HostapConfig.

TEST=Ran:
./run_remote_tests.sh --remote=chromeos1-shelf1-host4 check_dtimperiod
./run_remote_tests.sh --remote=chromeos1-shelf1-host4 \
    network_WiFiMatFunc --args="test_pat=007*"
Which test both that the migrated test works and that the power save
migration is functionally identical to before.
BUG=chromium:230660

Change-Id: I4593bb6c61625ef69e0bf0256288a0cc9b031846
Reviewed-on: https://gerrit.chromium.org/gerrit/50065
Tested-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>