firmware_TouchMTB: new logs with raw data for unit tests
Re-design the unit tests for the summary modules for the purpose of a
more systematic coverage on different fw versions, different gestures,
different validators, single/multiple iterations, etc.

The raw data files are also kept so that they could be replayed
anytime when new logs need to be generated due to the change of log
format.

BUG=chromium:238039
TEST=On a chromebook
$ cd /usr/local/autotest/tests/firmware_TouchMTB
Make sure that the following unit test pass.
$ python tests/firmware_summary_unittest.py

Change-Id: Ie42ce42ac36d84bdf9e439cec5a48488f0c8b674
Reviewed-on: https://gerrit.chromium.org/gerrit/50311
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>