autotest: Remove chaos lab packet capture CLI utility
Originally, I had written this to enable us to take packet captures
while developers tweaked things on a DUT by hand.  This made sense when
the AP's in the lab were always on and ready to be associated with.
However, in the new style of tests, AP's are off until asked for, and
need to be configured before use.  Since this script doesn't handle
configuring an AP, it is not very useful any more.

TEST=None.
BUG=chromium:238868

Change-Id: Iaa87887185ec923934b7093a890893c3d46b1bc4
Reviewed-on: https://gerrit.chromium.org/gerrit/50382
Tested-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Jason Abele <jabele@chromium.org>
Reviewed-by: Tan Gao <tgao@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>