autotest (shill): add test for DHCP timeout handling
We found that shill crashes if a DHCP timeout is encountered
on an Ethernet interface. To reproduce the crash, as well
as to guard against regressions, add an autotest that exercises
the same code path.

While there: modify dhcp_test_base.py to not modify TestFail
exceptions. This makes the failure output more consistent
with other test scripts.

BUG=chromium:237576
TEST=manual

Manual test
-----------
- test with shill at revision 1370986f -- FAILS
- test with shill and CL:50509 -- PASSES

Change-Id: I015b1a7d9edd798344a838668416adb963c65c29
Reviewed-on: https://gerrit.chromium.org/gerrit/50501
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>