autotest: Migrate network_WiFiMatFunc/020CheckWMM to new framework
Add support for forcing WMM on/off on a router.  Fix a problem with
parsing ping arguments, and fix some style issues with quotation marks.
Be more strict about parsing QoS parameters to ping, since this is the
only place where we use this parameter, and it does so correctly.  We
can safely assert that all new uses of this parameter should be correct
or the proper usage updated.

TEST=This is a test, it passes on snow and lumpy.
BUG=chromium:230660

Change-Id: Idf4b75591b2af0cacdaa8ddff1e718b358e6be25
Reviewed-on: https://gerrit.chromium.org/gerrit/51198
Tested-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>