firmware_TouchMTB: introduce a geometry package
The new requirement of the drumroll validator is that all points
from the same finger should be within 2 circles of radius X mm. We
could classify the points of the same finger into two clusters with
some method. For each cluster, calculate the minimal enclosing circle
of the points. As a first step, this patch implements an algorithm
proposed by Emo Welzl which is reasonably fast to compute such a
minimal enclosing circle.

In this patch, we introduce a geometry package which includes an
elements module to handle basic shapes, and a minicircle module
to calculate the minimal enclosing circle of a set of points.

We use the enclosingCircle funciton in octave's geometry package
to verify the correctness of the related unit tests.

BUG=chromium:241092
TEST=On a chromebook
$ cd /usr/local/autotest/tests/firmware_TouchMTB
Make sure that the following unit test pass.
$ python tests/geometry_unittest.py

Change-Id: I586d8dff6e825502e88873a951230e1335b19f99
Reviewed-on: https://gerrit.chromium.org/gerrit/51300
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>