Add robustness to the port creation of the devserver.
The devserver was really never intended to be rapidly created/destroyed.
Given that this may happen (and that we are doing this more and more as
we scale), we should add robustness around devserver creation because
the port might still be being cleaned up while we are starting a new
omaha server.

This CL just loops around the devserver creation component and cleans
up a little of the code around it.

BUG=chromium:234725
TEST=pylint,pyflakes,running test now.

Change-Id: I6df57222f4fffd1f02e7c7f3fd0b4a5253f7a5f9
Reviewed-on: https://gerrit.chromium.org/gerrit/51327
Commit-Queue: Chris Sosa <sosa@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>
Tested-by: Chris Sosa <sosa@chromium.org>