[autotest] Server jobs include the hostname in their job_status.
Three attributes are essential to create a link to the logs
of a failure, the job_id, the owner and hostname. This change
includes the hostname in the status returned when a server job
fails. The status will always have the hostname attribute,
and even if it's set to None the reporter won't fail, so we
can blidly include it in the result we return to dynamic_suite.
This only affects server jobs that fail, not ones that Abort.

TEST=Raised an exception in the server job and checked that
     the bug filed includes a link to the log.
BUG=None

Change-Id: I23fa5eadcf6d969a923159255c62ace9d5f8432b
Reviewed-on: https://gerrit.chromium.org/gerrit/51352
Tested-by: Prashanth Balasubramanian <beeps@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Queue: Prashanth Balasubramanian <beeps@chromium.org>