firmware_TouchMTB: get two farthest clusters
We add a function to compute the two farthest clusters of a given
list of points. First of all, the two farthest points are derived
which represent the two clusters. Then classify every point to the
nearer cluster.

BUG=chromium:241092
TEST=On a chromebook
$ cd /usr/local/autotest/tests/firmware_TouchMTB
Make sure that the following unit test pass.
$ python tests/simple_geometry_unittest.py

Change-Id: Id7881b19791a0e41a5c1ed1e2d9279ad48ccd3cf
Reviewed-on: https://gerrit.chromium.org/gerrit/51443
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>