[autotest] Bug reporter uses apiary backend to read issue metadata.
This cl executes the first part of the bug reporters transition to the new
apiary api. To read anything from the tracker we use phapi_lib (the apiary
backend) and to write anything to the tracker we use gdata_lib. In time we
will transition completely to using phapi_lib.

TEST=Deduped Restrict-View-Google bugs with hard to find anchors in their
     summaries and/or comments.
BUG=chromium:241510

Change-Id: Ib35c9ebd9a8f72c2e73e0280f8e6b6421ace2b8a
Reviewed-on: https://gerrit.chromium.org/gerrit/56899
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Queue: Prashanth Balasubramanian <beeps@chromium.org>
Tested-by: Prashanth Balasubramanian <beeps@chromium.org>