[autotest] add an ignore_deps parameter to Suite constructor
This CL adds a ignore_deps flag to the Suite constructor. When the flag
is passed in as true, jobs will be scheduled without using any of the
dependencies specified in their control file's DEPENCENCY attribute.

This flag is useful in test_that, to avoid the need to add label objects
and assign labels to the (one-and-only) host that will exist in its
sqlite in-memory afe database.

BUG=chromium:236471
TEST=Modified suite unit tests to be sensitive to dependencies.
Added a test that uses the ignore_deps flag. All existing and new tests
pass.

Change-Id: I4a0de64fcc9e197f357106279fd8da35ab8ba90d
Reviewed-on: https://gerrit.chromium.org/gerrit/55855
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Tested-by: Aviv Keshet <akeshet@chromium.org>