Restore hack due to build number issues.
Release trybot images may have a -b suffix which confuses the heck
out of full_release_test when finding the source image payload.
I had this in before but removed it when I refactored this code (note
even the comment was leftover).

This code simply says if the version I'm finding the source image for
is a superset of the release, treat the archive_url as the whole
archive_url to find the source payload.

BUG=chromium:240907
TEST=Ran it against trybots, pgo-use and release images.

Change-Id: Icc3f9c344664a6732f297d43774fc6a80942e171
Reviewed-on: https://gerrit.chromium.org/gerrit/55599
Tested-by: Chris Sosa <sosa@chromium.org>
Reviewed-by: Alex Miller <milleral@chromium.org>
Commit-Queue: Chris Sosa <sosa@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>