power_suspend: Use lock_vt where system does too
Use the lock_vt functionality on systems where the regular suspend/resume
process does the same. This will ensure that we actually test the same
code paths that users are executing.

It would have been nice to pass in the lockvt USE flag all the way into
the testcase, since that is how we specify it per-board, but it seems
to be quite awkward to do all the way down to the library. So I'm doing
the second best for now, and adding a small dictionary for the platforms
that should not enable it.

BUG=chromium:241491
TEST=ran platform_SuspendStress for 20 iterations without seeing lockups
due to VT switching

Change-Id: Iff3e91695f2beccdffb21692d062731dcab5d40c
Signed-off-by: Olof Johansson <olofj@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/55817
Reviewed-by: Daniel Erat <derat@chromium.org>