firmware_TouchMTB: add MtbStateMachine class in mtb module
In this patch, we use MtbStateMachine class to emulate the internal
state machine in the kernel driver. This simplifies the way of
extracting points from packets, and also improves the readability.

Besides, the methods of determining MTB events are moved into its
own class. This makes the mtb module cleaner.

BUG=chromium:241092
TEST=On a chromebook
$ cd /usr/local/autotest/tests/firmware_TouchMTB
Make sure that the following unit test pass.
$ python tests/mtb_unittest.py

Change-Id: I141f3c781326a4020f2c5e7fb7cdf6554f733986
Reviewed-on: https://gerrit.chromium.org/gerrit/56410
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>