cellular: Fix network_3GDisconnectFailure.3GPP auto-connect issue.
Since a 3GPP service is set to auto-connect by default, the modem gets
connected before the initial call to |network.ResetAllModems|. As this
explicitly disables the modem, and a disable will by default try to
Disconnect, however, since the pseudomodems used in this test fail to
disconnect, the exception raised propagate to the test and cause a test
failure.

This CL addresses this by making the custom Disconnect implementations
not fail if they were initiated by a Disable operation.

BUG=chromium:242648
TEST=Running network_3GDisconnectFailure should pass.

Change-Id: I359c03c419dac11a2a1c583adfd19216a6db05f8
Reviewed-on: https://gerrit.chromium.org/gerrit/56098
Commit-Queue: Arman Uguray <armansito@chromium.org>
Reviewed-by: Arman Uguray <armansito@chromium.org>
Tested-by: Arman Uguray <armansito@chromium.org>