[autotest] Enum get_value works when passed a long
Prior to this CL:
> import common
> from autotest_lib.client.common_lib import control_data
> control_data.CONTROL_TYPE.get_value('client')
2
> control_data.CONTROL_TYPE.get_value(2)
2
> control_data.CONTROL_TYPE.get_value(2L)
AttributeError: 'long' object has to attribute 'upper'

After this CL:
[snip]
> control_data.CONTROL_TYPE.get_value(2L)
2L

The reason this matters is that, occasionally, enums stored in one of
our databases as an integer are being stored as a long. Hardening Enum
so it sanely handles long ints removes the need to sanity check whether
enums coming out of the db are ints or longs.

BUG=See above
TEST=unit tests pass; Also see above.

Change-Id: I3ae84e8593daa559f530c99c6454eb2336b51303
Reviewed-on: https://gerrit.chromium.org/gerrit/56116
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Tested-by: Aviv Keshet <akeshet@chromium.org>
Commit-Queue: Aviv Keshet <akeshet@chromium.org>