Add new test that has shorter reboot/down timeouts and increase others.
This CL adds platform_RebootAfterUpdate which specifically tests shorter
reboot timeouts. I originally planned on setting down_timeout and
reboot_timeout and passing these to reboot but it seemed to be actually a
worse solution for 2 reasons:

1) Since these are client errors, they are reinterpreted as AutoservRunErrors.
I'd literally have to parse the ssh commands to figure whether it was associated
with either not shutting down in time or rebooting in time.
2) Even with (1) wait_for_reboot in client/common_lib/hosts/base_classes.py
waits using the down_timeout + reboot_timeout -- sort of. It doesn't subtract
the time used waiting for wait_down from the total time in wait_up. This is
pretty bad logic and isn't really that useful to set a good timeout.

Given these, it doesn't make much sense to separate down/reboot errors (which I
originally planned to do). So all I could do was sort of pass this one aggregate
value (which wait_for_reboot doesn't allow). So given all these, I decided that
allowing the long timeouts and measuring the time was the best solution for
our problem.

BUG=chromium:242384
TEST=Ran it -- manually changed time outs to see failures correctly.

Change-Id: I52103c8f835cea326bf7efbeb126d9f145988a43
Reviewed-on: https://gerrit.chromium.org/gerrit/56348
Reviewed-by: Chris Sosa <sosa@chromium.org>
Tested-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Chris Sosa <sosa@chromium.org>