power_suspend: Fix LOCKVT mechanism for different suspend methods
CL:55817 introduced a mechanism to run the lock_vt helper before suspend
on certain boards. In order to do that, it modified the signature of the
Suspender._suspend variable function call, but only adapted one of the
possible functions that could be assigned to that variable (thus
breaking tests such as power_SuspendStress.idle that use other suspend
methods).

This patch moves the execution of the lock_vt helper into the
Suspender class, so that the mechanism is isolated within the
power_suspend.py file and will work with any suspend method that is
passed to it. (Since some suspend methods run through powerd, this may
cause lock_vt to be called twice, which should not be a problem.)

BUG=None
TEST=Manual (ran power_Resume and power_SuspendStress.idlePerBuild on a
Snow and a Lumpy).

Change-Id: I110768729da7f79062c1f3900ae6d6d233760c62
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/56530
Reviewed-by: Mark Koudritsky <kamrik@chromium.org>